Silently introduce `set_psql_array` to workaround issue about array + Postgres support
This is a temporary fix on the public interface though (so we will keep it around probably deprecated later).

This enables `substreams init` to handle array for Postgres case. A similar method is available for clickhouse but it seems `substreams-sink-sql` is unable to handle it properly, at least for `Int256/Uint256` types.